kupica/src/main/java/org/nightdivers/kupica/controller/ViewController.java
package org.nightdivers.kupica.controller;

import lombok.RequiredArgsConstructor;
import org.nightdivers.kupica.service.AuthService;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;

@RequiredArgsConstructor
@RequestMapping("/")
@Controller
public class ViewController {

    private final AuthService authService;

    @GetMapping
    public String index() {
        return "index";
    }

    @GetMapping("/login")
    public String login() {
        return "login";
    }

    @GetMapping("/register")
    public String register() {
        return "register";
    }
}
